Reorganize the Connectors spec module (#7677)
