+++ apollo-federation/src/connectors/models.rs
@@ -26,10 +26,10 @@ use super::JSONSelection;
 use super::PathSelection;
 use super::id::ConnectorPosition;
 use super::json_selection::ExternalVarPaths;
-use super::spec::schema::ConnectBatchArguments;
-use super::spec::schema::ConnectDirectiveArguments;
-use super::spec::schema::ErrorsArguments;
-use super::spec::schema::SourceDirectiveArguments;
+use super::spec::connect::ConnectBatchArguments;
+use super::spec::connect::ConnectDirectiveArguments;
+use super::spec::errors::ErrorsArguments;
+use super::spec::source::SourceDirectiveArguments;
 use super::variable::Namespace;
 use super::variable::VariableReference;
 use crate::connectors::ConnectSpec;
