--- apollo-federation/src/connectors/spec/schema.rs
@@ -1,152 +1,0 @@
-use apollo_compiler::Name;
-use apollo_compiler::name;
-use http::Uri;
-
-use crate::connectors::ConnectorPosition;
-use crate::connectors::SourceName;
-use crate::connectors::json_selection::JSONSelection;
-use crate::connectors::models::Header;
-
-pub(crate) const CONNECT_DIRECTIVE_NAME_IN_SPEC: Name = name!("connect");
-pub(crate) const CONNECT_SOURCE_ARGUMENT_NAME: Name = name!("source");
-pub(crate) const CONNECT_SELECTION_ARGUMENT_NAME: Name = name!("selection");
-pub(crate) const CONNECT_ENTITY_ARGUMENT_NAME: Name = name!("entity");
-
-pub(crate) const CONNECT_HTTP_NAME_IN_SPEC: Name = name!("ConnectHTTP");
-pub(crate) const CONNECT_BATCH_NAME_IN_SPEC: Name = name!("ConnectBatch");
-pub(crate) const CONNECT_BODY_ARGUMENT_NAME: Name = name!("body");
-
-pub(crate) const ERRORS_NAME_IN_SPEC: Name = name!("ConnectorErrors");
-pub(crate) const ERRORS_ARGUMENT_NAME: Name = name!("errors");
-pub(crate) const ERRORS_MESSAGE_ARGUMENT_NAME: Name = name!("message");
-pub(crate) const ERRORS_EXTENSIONS_ARGUMENT_NAME: Name = name!("extensions");
-
-pub(crate) const SOURCE_DIRECTIVE_NAME_IN_SPEC: Name = name!("source");
-pub(crate) const SOURCE_NAME_ARGUMENT_NAME: Name = name!("name");
-pub(crate) const BATCH_ARGUMENT_NAME: Name = name!("batch");
-
-pub(crate) const SOURCE_HTTP_NAME_IN_SPEC: Name = name!("SourceHTTP");
-pub(crate) const SOURCE_BASE_URL_ARGUMENT_NAME: Name = name!("baseURL");
-pub(crate) const HTTP_ARGUMENT_NAME: Name = name!("http");
-pub(crate) const HEADERS_ARGUMENT_NAME: Name = name!("headers");
-
-pub(crate) const PATH_ARGUMENT_NAME: Name = name!("path");
-pub(crate) const QUERY_PARAMS_ARGUMENT_NAME: Name = name!("queryParams");
-
-pub(crate) const HTTP_HEADER_MAPPING_NAME_IN_SPEC: Name = name!("HTTPHeaderMapping");
-pub(crate) const HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME: Name = name!("name");
-pub(crate) const HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME: Name = name!("from");
-pub(crate) const HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME: Name = name!("value");
-
-pub(crate) const JSON_SELECTION_SCALAR_NAME: Name = name!("JSONSelection");
-pub(crate) const URL_PATH_TEMPLATE_SCALAR_NAME: Name = name!("URLTemplate");
-
-/// Arguments to the `@source` directive
-#[cfg_attr(test, derive(Debug))]
-pub(crate) struct SourceDirectiveArguments {
-    /// The friendly name of this source for use in `@connect` directives
-    pub(crate) name: SourceName,
-
-    /// Common HTTP options
-    pub(crate) http: SourceHTTPArguments,
-
-    /// Configure the error mapping functionality for this source
-    pub(crate) errors: Option<ErrorsArguments>,
-}
-
-/// Common HTTP options for a connector `@source`
-#[cfg_attr(test, derive(Debug))]
-pub struct SourceHTTPArguments {
-    /// The base URL containing all sub API endpoints
-    pub(crate) base_url: Uri,
-
-    /// HTTP headers used when requesting resources from the upstream source.
-    /// Can be overridden by name with headers in a @connect directive.
-    pub(crate) headers: Vec<Header>,
-    pub(crate) path: Option<JSONSelection>,
-    pub(crate) query_params: Option<JSONSelection>,
-}
-
-/// Configure the error mapping functionality for a source or connect
-#[cfg_attr(test, derive(Debug))]
-pub(crate) struct ErrorsArguments {
-    /// Configure the mapping for the "message" portion of an error
-    pub(crate) message: Option<JSONSelection>,
-
-    /// Configure the mapping for the "extensions" portion of an error
-    pub(crate) extensions: Option<JSONSelection>,
-}
-
-/// Arguments to the `@connect` directive
-///
-/// Refer to [ConnectSpecDefinition] for more info.
-#[cfg_attr(test, derive(Debug))]
-pub(crate) struct ConnectDirectiveArguments {
-    pub(crate) position: ConnectorPosition,
-
-    /// The upstream source for shared connector configuration.
-    ///
-    /// Must match the `name` argument of a @source directive in this schema.
-    pub(crate) source: Option<SourceName>,
-
-    /// HTTP options for this connector
-    ///
-    /// Marked as optional in the GraphQL schema to allow for future transports,
-    /// but is currently required.
-    pub(crate) http: Option<ConnectHTTPArguments>,
-
-    /// Fields to extract from the upstream JSON response.
-    ///
-    /// Uses the JSONSelection syntax to define a mapping of connector response to
-    /// GraphQL schema.
-    pub(crate) selection: JSONSelection,
-
-    /// Entity resolver marker
-    ///
-    /// Marks this connector as a canonical resolver for an entity (uniquely
-    /// identified domain model.) If true, the connector must be defined on a field
-    /// of the Query type.
-    pub(crate) entity: bool,
-
-    /// Settings for the connector when it is doing a $batch entity resolver
-    pub(crate) batch: Option<ConnectBatchArguments>,
-
-    /// Configure the error mapping functionality for this connect
-    pub(crate) errors: Option<ErrorsArguments>,
-}
-
-/// The HTTP arguments needed for a connect request
-#[cfg_attr(test, derive(Debug))]
-pub struct ConnectHTTPArguments {
-    pub(crate) get: Option<String>,
-    pub(crate) post: Option<String>,
-    pub(crate) patch: Option<String>,
-    pub(crate) put: Option<String>,
-    pub(crate) delete: Option<String>,
-
-    /// Request body
-    ///
-    /// Define a request body using JSONSelection. Selections can include values from
-    /// field arguments using `$args.argName` and from fields on the parent type using
-    /// `$this.fieldName`.
-    pub(crate) body: Option<JSONSelection>,
-
-    /// Configuration for headers to attach to the request.
-    ///
-    /// Overrides headers from the associated @source by name.
-    pub(crate) headers: Vec<Header>,
-
-    /// A [`JSONSelection`] that should resolve to an array of strings to append to the path.
-    pub(crate) path: Option<JSONSelection>,
-    /// A [`JSONSelection`] that should resolve to an object to convert to query params.
-    pub(crate) query_params: Option<JSONSelection>,
-}
-
-/// Settings for the connector when it is doing a $batch entity resolver
-#[derive(Clone, Copy, Debug)]
-pub struct ConnectBatchArguments {
-    /// Set a maximum number of requests to be batched together.
-    ///
-    /// Over this maximum, will be split into multiple batch requests of `max_size`.
-    pub max_size: Option<usize>,
-}
