+++ apollo-federation/src/connectors/validation/connect.rs
@@ -20,8 +20,8 @@ use crate::connectors::Namespace;
 use crate::connectors::SourceName;
 use crate::connectors::id::ConnectedElement;
 use crate::connectors::id::ObjectCategory;
-use crate::connectors::spec::schema::CONNECT_SOURCE_ARGUMENT_NAME;
-use crate::connectors::spec::schema::SOURCE_NAME_ARGUMENT_NAME;
+use crate::connectors::spec::connect::CONNECT_SOURCE_ARGUMENT_NAME;
+use crate::connectors::spec::source::SOURCE_NAME_ARGUMENT_NAME;
 use crate::connectors::validation::connect::http::Http;
 use crate::connectors::validation::errors::Errors;
 use crate::connectors::validation::graphql::SchemaInfo;
