+++ apollo-federation/src/connectors/validation/connect/http.rs
@@ -12,9 +12,9 @@ use shape::Shape;
 use crate::connectors::HTTPMethod;
 use crate::connectors::Namespace;
 use crate::connectors::SourceName;
-use crate::connectors::spec::schema::CONNECT_BODY_ARGUMENT_NAME;
-use crate::connectors::spec::schema::CONNECT_SOURCE_ARGUMENT_NAME;
-use crate::connectors::spec::schema::HTTP_ARGUMENT_NAME;
+use crate::connectors::spec::connect::CONNECT_BODY_ARGUMENT_NAME;
+use crate::connectors::spec::connect::CONNECT_SOURCE_ARGUMENT_NAME;
+use crate::connectors::spec::http::HTTP_ARGUMENT_NAME;
 use crate::connectors::string_template;
 use crate::connectors::string_template::Part;
 use crate::connectors::string_template::StringTemplate;
