+++ apollo-federation/src/connectors/validation/errors.rs
@@ -16,9 +16,9 @@ use super::expression::MappingArgument;
 use super::expression::parse_mapping_argument;
 use crate::connectors::JSONSelection;
 use crate::connectors::Namespace;
-use crate::connectors::spec::schema::ERRORS_ARGUMENT_NAME;
-use crate::connectors::spec::schema::ERRORS_EXTENSIONS_ARGUMENT_NAME;
-use crate::connectors::spec::schema::ERRORS_MESSAGE_ARGUMENT_NAME;
+use crate::connectors::spec::errors::ERRORS_ARGUMENT_NAME;
+use crate::connectors::spec::errors::ERRORS_EXTENSIONS_ARGUMENT_NAME;
+use crate::connectors::spec::errors::ERRORS_MESSAGE_ARGUMENT_NAME;
 use crate::connectors::string_template::Expression;
 use crate::connectors::validation::Code;
 use crate::connectors::validation::Message;
