+++ apollo-federation/src/connectors/validation/source.rs
@@ -8,9 +8,9 @@ use super::coordinates::SourceDirectiveCoordinate;
 use super::errors::ErrorsCoordinate;
 use super::http::UrlProperties;
 use crate::connectors::SourceName;
-use crate::connectors::spec::schema::HTTP_ARGUMENT_NAME;
-use crate::connectors::spec::schema::SOURCE_BASE_URL_ARGUMENT_NAME;
-use crate::connectors::spec::schema::SOURCE_NAME_ARGUMENT_NAME;
+use crate::connectors::spec::http::HTTP_ARGUMENT_NAME;
+use crate::connectors::spec::source::SOURCE_BASE_URL_ARGUMENT_NAME;
+use crate::connectors::spec::source::SOURCE_NAME_ARGUMENT_NAME;
 use crate::connectors::validation::Code;
 use crate::connectors::validation::Message;
 use crate::connectors::validation::coordinates::BaseUrlCoordinate;
